backend/app/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "app"
dynamic = ["version"]
description = ""
readme = "README.md"
requires-python = ">=3.12,<4.0"
license = "AGPL-3.0-or-later"
keywords = []
authors = [
  { name = "U.N. Owen", email = "info@example.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.12",
]
dependencies = [
  "inboard[fastapi]==0.72.2",
  "python-multipart>=0.0.12,<1.0.0",
  "email-validator>=2.2.0",
  "requests>=2.32.3",
  "requests-http-signature>=0.7.1",
  "celery>=5.4.0",
  "argon2-cffi>=23.1.0",
  "cryptography>=43.0.1",
  "passlib[argon2]>=1.7.4,<2.0.0",
  "tenacity>=9.0.0",
  "pydantic>=2.11.4",
  "pydantic-settings>=2.6.0",
  "python-ulid[pydantic]>=3.0.0",
  "emails>=0.6.0",
  "sentry-sdk[fastapi]>=1.40.6,<2.0.0",
  "jinja2>=3.1.4",
  "alembic>=1.13.3",
  "sqlalchemy>=2.0.36",
  "pyjwt>=2.9.0",
  "httpx>=0.27.2",
  "psycopg[binary]>=3.2.3",
  "setuptools>=75.2.0",
  "sqlalchemy-utils>=0.41.2",
  "orjson>=3.10.7",
  "boto3>=1.35.43",
  "Babel>=2.16.0",
  "fastapi-cache2[redis]>=0.2.2",
  "websockets>=13.1",
  "ip2location>=8.10.4",
  "tomli-w>=1.2.0",
  "bovine>=0.5.18",
  "faststream[rabbit]>=0.5.40",
  "regex>=2024.11.6",
  "nh3>=0.2.21",
  "Faker>=37.3.0",
  "mistune>=3.1.3"
]

[project.optional-dependencies]
checks = [
  "black>=24.10.0",
  "mypy>=1.12.0",
  "isort>=5.13.2",
  "autoflake>=2.3.1",
  "flake8>=7.1.1",
]

[project.urls]
Documentation = "https://github.com/unknown/app#readme"
Issues = "https://github.com/unknown/app/issues"
Source = "https://github.com/unknown/app"

[tool.hatch.version]
path = "app/__version__.py"

# [dirs.env]
# virtual = "./.venv"

[tool.hatch.envs.default]
python="3.12"  # <--
installer = "uv"
dev-mode = true

[tool.hatch.build.targets.sdist]
include = ["/app"]

[tool.hatch.envs.production]
dev-mode = false
features = []
path = ".venv"
dependencies = []

[tool.hatch.envs.development]
dev-mode = true
features = []
path = ".venv"
dependencies = [
  "pytest>=8.3.3",
  "jupyterlab>=4.3.5"
]

[tool.hatch.envs.lint]
detached = true
dependencies = [
  "black>=24.10.0",
  "mypy>=1.12.0",
  "isort>=5.13.2",
]
[tool.hatch.envs.lint.scripts]
style = [
  "isort --check --diff {args:.}",
  "black --check --diff {args:.}",
]
fmt = [
  "black {args:.}",
  "isort {args:.}",
  "style",
]
all = [
  "style",
  "typing",
]

[tool.black]
target-version = ["py312"]
line-length = 120

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 120
src_paths = ["app", "tests"]

[tool.mypy]
files = ["**/*.py"]
plugins = "pydantic.mypy"
show_error_codes = true
strict = true
